# Improve performance of array predicates